Wrap class in a block
As of Rakudo 2015.05, the `unit` declarator is required before using
`module`, `class` or `grammar` blockless declarations.  An alternative is to
wrap the class, role, grammar or module in a block, as done in this change,
since it fits the code's current visual structure.

Code still using the old blockless semicolon form will throw a warning. This
commit stops the warning from appearing in the new Rakudo.